Strip system paths from RPATH
Several Linux distributions forbid setting system library paths into RPATH.
See discussion about this issue at
  https://groups.google.com/group/oce-dev/browse_thread/thread/e0fbdeb52b90111b
and in particular
  https://groups.google.com/group/oce-dev/msg/dff0e5ac3eb67fb7
  http://public.kitware.com/Bug/view.php?id=11751#c25029

It seems from the bugreport above that INSTALL_RPATH_USE_LINK_PATH
might be the solution, but I have been unable to provide a working
configuration.

This commit is a slightly different version of a patch provided by
Richard M. Shaw:
  https://github.com/tpaviot/oce/pull/199

A new option OCE_RPATH_FILTER_SYSTEM_PATHS is added to discard this
behavior.  According to documentation, the determination of system
paths is empirical and may be wrong, this option allows to not rely
on it.